feat: Consolidate NotifiSubscriptionContextProvider and NotifiFormProvider
