+++ packages/notifi-react-example/src/NotifiContextWrapper/index.ts
@@ -1,0 +1,1 @@
+export * from './SolanaNotifiContextWrapper';
